Enabling orders to be created and updated with nested attributes and adding commit action to check stock and set order status.

+++ db/migrate/20220930132300_create_products.rb
@@ -1,7 +1,7 @@
 class CreateProducts < ActiveRecord::Migration[6.1]
   def change
     create_table :products do |t|
-      t.decimal :price, default: 100
+      t.integer :price, default: 100
       t.string :type
       t.string :name
